Remove twemoji
..because it was causing #543 and other problems.

Basically there is an exception when twemoji is parsing the dom tree.
Some text node do not have a .parentElement which creates exceptions.
I also tried using twemoji.parse based on string but
that would turn pure text into text+images and i am
not sure how to handle that with trifl, yet.
Why is twemoji needed on windows anyway ?
because noto font does not work ?